+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Added
+
+- Add `cluster-admin` flag to `login` command, which allows full access for GiantSwarm staff.
+
 ## [0.13.0] - 2020-11-20
 
 ### Removed